stp2/construction_crm/requirements.txt
Django==4.2.7
django-crispy-forms==2.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.5
psycopg2-binary==2.9.7
Pillow==10.0.1
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
